dtrace: Simplify the cleanup process. Fix a panic.
This commit simplifies the way HyperTrace probes are cleaned up. Prior to this
commit, dtrace_nvmids could grow to a very large number and end up in a
situation where dtrace_vprobes[vmid] was NULL when destroying vprobes. This
would then panic the kernel because we weren't checking for it.

However, dtrace_nvmids was a broken mechanism to begin with, because we didn't
know which vmids were active and which weren't anyway. Instead of doing it based
on counting which vmids are active and which aren't, just attempt to clean them
all up and perform the NULL checks in places that make sense.