dtraced: Wait 5 seconds for the child to respond, or kill it.
This is a simple heuristic that prevents a deadlock from happening when
dtrace(1) has a bug. This shouldn't be needed if dtrace(1) always responds, but
if it hangs somewhere, we don't really want it to deadlock dtraced. Instead, 5
seconds we currently deem to be plenty of time for dtrace(1) to respond to us,
otherwise we SIGKILL it and move on.