dtraced: Create a cleanup action in dtraced.
This is activated by `dtracedctl --cleanup-state` currently. It's mostly a
workaround for a race condition which happens when dtraced forks a process
in the guest, but doesn't get the time to respond to the host dtrace(1)
process with its pid.

This results in the guest continuing to run a process that no longer has
anything listening for its DTrace state on the host, or a way to kill it.
As a workaround currently, we allow the user to simply run `dtracedctl` to
address various issues like this.