dtraced: Re-exec on cleanup.
We previously tried to do a lot of the manual cleanup, but if the state we
are in is unrecoverable, we can't really do much. Instead of doing that, we
simply kill all the known processes and clean up the joblist, destroy our
state (which also sends SIGTERM to all threads, but that will be caught by
our signal handler setting state.shutdown) and call execve to re-exec dtraced
with the arguments it was invoked with.