dtraced: Infer if we should run in minion or overlord mode.
Previously, we required the user to always specify '-O' or '-m' to dtraced.
In this commit, guided by #128 we now infer the mode we should run in.

Currently, the idea is to run in overlord mode in any situation other than
when dtraced is run inside a bhyve guest.

Also fix the help page as '-d' now stands for "debug" rather than "daemonize".

Closes #130.