dtrace: Process printf in the same way as a DIFEXPR.
This is needed, as a printf action can contain a DIFO which we need to further
link. An example script that would cause this behaviour to fail (if we didn't
process DTRACEACT_PRINTF as a DTRACEACT_DIFEXPR) is:

```
dtrace -E -n 'vm*:syscall:::entry { printf("hello world %s", curthread->td_proc->p_comm); }'
```

We would end up with a return size of 0 for the action inside the printf, and
nothing would be printed.

This fully solves #139, which was previously partially resolved when fixing
array returns.