dtrace: Relocate PUSHTV types if necessary.
Previously, we did not relocate PUSHTV types if they were needed (e.g. on a
subroutine call) which caused a failure in the kernel verifier, because we were
not pushing a CTF type onto the stack, but rather a BOTTOM type.

This commit fixes it and ensures that the guest correctly relocates PUSHTV.