dtrace: Relocate typecast correctly.
Previously, we didn't account for unneeded sign-extension in a typecast. We
generate it every time because we don't know if it will be needed. However,
we know for a fact that it is not needed for uintptr_t, and in fact it destroys
the pointer.

This commit ensures that we don't do any form of sign extension on uintptr_t
and is the final in the series of commits that resolves a blocking issue for