dtrace: Start handling user addresses.
This commit implements the host side of handling user addresses in
the system call interface. Many system calls take userspace
addresses which are later copied in through copyin() in the kernel.

However, these addresses are passed as user addresses to DTrace in
the system call interface using the systrace provider. The user then
might reasonably write a script such as:

dtrace -n 'sycal::open:entry { trace(copyinstr(arg0)); }'

which would take a path name address, which is a userspace address
in arg0 and then copy it into DTrace local (kernel) memory to then
access it.

However, in HyperTrace we simply don't know how to translate this
guest user address to the guest host address, as we do not have access
to guest's page tables. Instead, we currenly assume that HyperTrace
will ensure that the guest performs the translation before it hands off
arg0 to the host DTrace instance, and the address we get will be a
kernel address. This doesn't currently happen, so this doesn't work and
an appropriate FIXME is added.

This commit is the first step (host side) of fixing #142.