dtrace: Implement some missing parts of immstack().
This commit implements dumping of immstack offsets,
adds a license and makes it keyable for aggregations.

Furthermore, we implement a new action, PRINTIMMSTACK
which is generated by the compiler and can't directly
be used by the script writer. This is going to act as
a mechanism to synchronize the guest and host once we
have a compiler pass implementing the partial compute.

Finally, exclude things that recursed into DTrace as a
result of immstack() from fbt and add a license.