dtrace: Implement the compiler bits for immstack().
This commit implements the necessary compiler transformations in
order to get immstack() to work in the guest and correctly be printed
on the host. This gives us a way to inspect guest stack traces
with HyperTrace.

However, this is currently extremely poor performing due to a symbol
table walk + linker file walk in ddb. We will need to optimize this
if the performance becomes unsatisfactory. Perhaps caching things could
work if done at the right place.

Closes #125.