dtrace: Fix a race condition in communication with dtraced.
This commit fixes a race condition which appears when we have multiple
VMs responding with their programs, which is then enqueued to be
processed by dtraced before the dtrace command line tool is able to
request a READ_DATA to dtraced. We then end up in a recv() in both
dtraced worker threads and the dtrace command line tool, leading to
a deadlock where we need to kill -9 dtraced.

This relates to #147.

This commit also fixes use of uninitialized memory in dtraced.