dtraced: Support human-readable client identification.
This commit adds a new field in the file descriptor list in dtraced.
This field (called 'ident') is used for clients to identify themselves
which can then be used in dtracedctl --show-stats and various other
useful things.

It is intended for clients to identify themselves in the initial
message they send once they connect to dtraced, and their
identification is not mandatory, but it is preferred.