dtrace: Fix a race condition in the immstack() cache.
This race condition is caused by a hash collision across multiple
CPUs. While it wouldn't crash the system, it would give the wrong
results when calling immstack() -- as it is possible that the pc
storing a particular symbol is overwritten, and therefore the cache
contains the wrong symbol for a particular pc.