dtrace: Fix a crash that happens when using immstack().
This crash seems to have been caused by the indirection in the linker
implementation in FreeBSD when called from the DTrace probe context
with interrupts masked. When LINKER_SEARCH_SYMBOL is called and the
underlying function pointer is dereferenced, it seems to sometimes
point to nonsensical code, or somehow end up in it.

Observed crashes include a page fault that wasn't handled by the
DTrace trap check function even though NOFAULT was set, a GP fault,
various timeouts when shutting down all logical CPUs, panics in
unrelated parts of the kernel, etc.

This commit simply inlines all of the ddb stuff into DTrace, thus
making it possible to instrument ddb again. Furthermore, it removes
all of the calls to printf, strcmp, etc from libkern which ensures
that those don't end up recursing into DTrace if instrumented in
some edge conditions.

In addition to that, this commit adds a new DTrace helper function
dtrace_strcmp() which is essentially just an implementation of strcmp
intended to be called from the probe context.

There is still a bug where any symbol from a kernel module is not
correctly resolved and therefore we print "??+offset" instead of
the symbol itself. This will be resolved in the follow-up commits.