dtrace: Set arg0-arg9 to be int, initial support for ldga.
This commit adheres more to the DTrace idea that arg0 through arg9 are
integers, while args[] array is the actual typed arguments as seen in C
code. We currently don't check for types defined in our D and C containers
and instead just support module and kernel types.

This is the first commit in a series of commits that aims to address
bugs #154 and #155.