dtrace: Fix relocation generation.
Previously, we generated relocations if we could show that the type
we are trying to process is not in the D CTF container. Initially, the
thought process was that the only thing that would be in the D CTF
container would be types that were defined by the D script that we are
currently processing, and therefore there were no relocations to apply.

However, libdtrace will copy types from other CTF containers into the
D CTF container under certain conditions, which causes us to not generate
relocations for types we know nothing about (e.g. `struct vnode`).

This commit introduces a fix for this particular issue.