dtrace: Correctly type-check strings in variable typechecking.
When type-checking variables, one of the tests ends up in a scenario
where we have a 'string' ctf type (as created by libdtrace in the D
container) compared to a D string type (DIF_TYPE_STRING). These two are
the same types, but we didn't really allow it to be treated that way
before.

This commit also adds a new typefile interface: dt_typefile_getctfp()
which returns the CTF pointer of a given typefile.