dtrace: Start carrying membinfo pointers with typecast.
This commit is in preparation for the further relocations that we
will apply in the linking stage in order to support relocating a
pattern such as:

usetx ...%r1
typecast ..., %r1
...
ret %r1

which currently ends up being skipped and we end up with a usetx
instruction in the linked program due to the typecast shadowing it
in our graph.

However, with this commit we will be able to correctly relocate the
usetx from the typecast instruction.

Additionally, this commit removes a lot of code which was used for
debugging purposes in order to simplify the codebase.