dtrace: Propagate type correctly on st* instructions.
A stb/sth/stw/stx instruction stores a value to the location referenced
by the source register. This means it doesn't actually update the type
of the value stored in that register.

As a result, we should simply preserve the type through the st*
instruction rather than overwriting it with whatever we were storing
to the location pointed to by the source register.

This commit also removes the assigning of type to "other" in the add
instruction, as we don't know that if other type is bottom that it
should become whatever it is we're adding to it.

Also fix alignment in an if statement.