net: Support packet-tagging through virtio and change how it works.
This was implemented before, but we seem to have lost the changes in one
of the merges when bhyve's virtio-net backend changed a fair bit. In this
commit we re-implement the changes that we have in light of the new
virtio-net backends supported in bhyve, and change what we carry with
packet tags.

Previously, we carried a `struct mbufid_info` which would contain the
actual mbufid and the information as to what is assigned (msgid + hostid).

With this commit, we drop that idea and simply carry around the `mbufid_t`
bit, without any information as to what is actually initialized.