dtrace: Do not warn about unstable probes for non-host tracing.
When running a DTrace script that is supposed to trace guests, we
were erroring out on probe arguments being referenced when they were
declared as unstable or didn't match types. We don't want to do this
as we aren't tracing the host machine. This commit simply adds a few
simple checks to ensure that we are only erroring out with this error
if we are going to trace the host.