dtrace: Only verify statements meant for the same targets.
When iterating through the statement list, we might start trying to
verify a statement for target A with the statement for target B. This
clearly can't work, so we should never attempt to do it.

VMs will answer with only their statements, so we should only really
verify their statements.