dtrace: Temporarily disable a check when cook() is called.
This is needed because 'cooking' isn't necessarily just type-checking
but also preparation and linking of arguments in internal libdtrace
data structures. This means that we can't just not call it for anything
that doesn't resolve to the current target, as we will be missing links
between nodes later on (e.g. associative arrays).