dtrace: elf: Correctly populate the action list.
Previously, we accidentally didn't populate the action list correctly
because we were assuming that prev wouldn't become NULL after the first
call, and we would have an action list. However, because of the function
exiting and resetting the value of 'prev', we need to actually preserve
this. We also need to account for failures and simply ignore them as the
"previous" action.

This commit simply keeps track of the last successfully allocated action
as the 'prev' action and ensures that we correctly build the action list
for further processing.