dtrace: elf: Fix handling of actions not meant for current target.
There were some bugs when it comes to memory management of actions that
were not made to run on the current target. They would get free'd and
then used later on as a result of the latest set of patches.

This commit ensures that any action in the list is always valid when
parsing ELF files and can safely be accessed.