dtrace: Simplify and correct how new probes are allocated.
Instead of trying to do legacy-style bitshifts that DTrace was using
to create vectors, we now simply use the id of the probe we're trying
to create and compute its next power of two. This means that we will
always be able to create the probe that we are trying to create.

The old approach had an issue where the new vector size was less than
the id of the probe we were trying to create, leading to a number of
potential panics.