dtrace: Squash multiple enablings into one for guest programs.
This commit creates a new function: dt_vprog_squash(). The function
takes all of the enablings of the same probe in a given program and
squashes them into a single enabling. We only expect two action kinds
to be present: DIFEXPR (hypercall + ret) and IMMSTACK. Because all
the predicates will happen on the host, this is fine to do.