dtrace: Correct how actions are handled in ELF files.
Firstly, we need to iterate from the ECB's action rather than the
statement's action when creating actions. This is because we want to
always create all of the actions, and the statement may not be the
first one in the ecbdesc.

Secondly, we were passing in &ad to our hashmap to save it as a key.
This obviously didn't work as ad was an argument to the function, so
passing in the address of a function argument to save for later use
is not going to work. This commit addresses that issue, removing a
double free that would manifest on the host and a verification issue.