dtrace: Fix a segfault caused by an incorrect assumption.
We were assuming that a statement's action can't be NULL, however a
simple script containing a single probe:

END
{
}

will create a statement without dtsd_action or dtsd_action_last being
set to non-NULL values. Therefore, this assumption was wrong and
as a result the above script would crash libdtrace.

This commit introduces code changes to ensure that the action now
can be NULL and no longer crashes libdtrace.