dtrace: hypertrace: Change the API to allow for errors & update code.
We need to check if a probe exists before we start dereferencing it.
Not doing this led to a panic caused by the following script:

END {}

This commit fixes the issue where we get a panic with these kinds of
scripts and updates the API of hypertrace to support returning of
errors (negative value).