dtrace: Simplify & correct the inprobe check to avoid recursion.
Prior to the recent patches, we needed to check for the vmid in order
to know if recursing into the ERROR probe is possible. Now, we can
recurse into any vmid's ERROR probe due to the way we attached the
VM's handle onto curthread rather than calling the function with it
as the argument.