dtraced: Overhaul the logging system.
This commit overhauls the logging system in a way that it creates macros for
each of the messages (error, debug, warning) as ERR, DEBUG, WARN. It also
ensures that every single error message has a file it occurred in as well as
the line in that file. This makes it easier to track what is going on when
debugging.